Internal: Rethink mutations writer (#731)
* internal: Simplify the `VirtualDOM` polling

* internal: Rethink the MutationsWriter impl of freya

* clippy

+++ crates/core/src/platform_state.rs
@@ -9,7 +9,7 @@ pub struct NativePlatformState {
     pub preferred_theme: PreferredTheme,
     pub navigation_mode: NavigationMode,
     pub information: PlatformInformation,
-    pub scale_factor: f32,
+    pub scale_factor: f64,
 }
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Default)]
